small cleanup and added a few missing validations

+++ backend/src/main/java/com/scottphebert/personalwebsite/repo/UserDetailsRepo.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import java.util.Optional;
 
 public interface UserDetailsRepo extends JpaRepository<UserDetails, Long> {
-    Optional<UserDetails> findById(int id);
+    Optional<UserDetails> findById(long id);
 }
